# docs/swagger.yaml
definitions:
  controllers.Response:
    properties:
      message: {}
    type: object
...
  models.TenderRequest:
    properties:
      budget:
        type: number
      client_id:
        type: integer
      deadline:
        type: string
      description:
        type: string
      file_url:
        type: string
      title:
        type: string
    required:
    - budget
    - client_id
    - deadline
    - description
    - title
    type: object
  models.Tenders:
    properties:
      budget:
        type: number
      client_id:
        type: integer
      created_at:
        type: string
      deadline:
        type: string
      description:
        type: string
      file_url:
        type: string
      id:
        type: integer
      status:
        type: boolean
      title:
        type: string
      updated_at:
        type: string
    required:
    - client_id
    type: object
...
info:
  contact:
    email: muhtorhongofurov@gmail.com
    name: Muxtorxon Gofurov
    url: https://github.com/muhtorkhon
  description: Tender Management Golang REST API
  title: Tender Management REST API
  version: "1.0"
paths:
...
  /tenders:
    get:
      description: Retrieve all tenders with pagination support
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Tenders'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.Response'
      summary: Get all tenders with pagination
      tags:
      - tender
    post:
      consumes:
      - application/json
      description: Creates a new tender with the provided details
      parameters:
      - description: Tender Request Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.TenderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Tenders'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.Response'
      security:
      - BearerAuth: []
      summary: Create a new tender
      tags:
      - tender
  /tenders/{client_id}:
    get:
      description: Retrieve all tenders for a specific client
      parameters:
      - description: Client ID
        in: path
        name: client_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tenders'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.Response'
      security:
      - BearerAuth: []
      summary: Get tenders by client ID
      tags:
      - tender
  /tenders/{id}:
    delete:
      consumes:
      - application/json
      description: Marks a tender as deleted by setting the `DeletedAt` timestamp.
      parameters:
      - description: Tender ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tender deleted successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/controllers.Response'
        "404":
          description: Tender not found or already deleted
          schema:
            $ref: '#/definitions/controllers.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.Response'
      security:
      - BearerAuth: []
      summary: Soft delete a tender by ID
      tags:
      - tender
    put:
      consumes:
      - application/json
      description: Updates the details of an existing tender
      parameters:
      - description: Tender ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Tender Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.TenderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.Response'
        "404":
          description: Tender not found
          schema:
            $ref: '#/definitions/controllers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.Response'
      security:
      - BearerAuth: []
      summary: Update an existing tender
      tags:
      - tender
  /tenders/restore/{id}:
    patch:
      consumes:
      - application/json
      description: Restores a tender by removing the `DeletedAt` timestamp, making
        it active again.
      parameters:
      - description: Tender ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tender restored successfully
          schema:
            type: string
        "404":
          description: Tender not found or already active
          schema:
            $ref: '#/definitions/controllers.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.Response'
      security:
      - BearerAuth: []
      summary: Restore a soft deleted tender by ID
      tags:
      - tender
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Bearer token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"